Added a full on Recipe API
Result - Holds the result IItemProvider, along side with count of the item. Optionally it can hold the group of the recipe, the path and custom prefixes/suffixes for the path and/or file name.

Input - Contains a maximum of 9 chars, representing 9 slots in the crafting GUI (3x3).

Grid - The Grid that will be used to create the layout of the recipe (3x3), **must have proportional recipe size**.

GridInput - Contains the combination of Grid and Input classes, where the chars specified in the Input will be used to represent the layout entries in the Grid (Grid chars must match with Input chars). Additionally the class contains a plethora of available methods to use for faster creation of layout/grid.

Now that that is out of the way, there is also CraftingRecipeMaker.
Main Functions: Utilizes the API and contains a big variety of methods that are used to generate the recipes automatically in `/generated/`.

And because in 1.16.3+ we utilize Smithing recipes, I also created SmithingRecipeMaker, it basically does what CraftingRecipeMaker does but for smithing recipes, in a simplified way.